+++ src/handlers/game.py
@@ -16,8 +16,7 @@ def handle(_, block: Block, __):
         for param in rule.description.split() }
     
     for param, line_number in params.items():
-        sieve = block.get_sieve()
-        (operand, values) = _get_operand_and_values(param, sieve, line_number)
+        (operand, values) = _get_operand_and_values(param, block.get_sieve(), line_number)
         block.upsert(operand, [ f'"{value}"' for value in values ])
 
     return block.get_raw_lines()