+++ src/ninja/__init__.py
@@ -1,3 +1,3 @@
 """Contains all the functionality used to comunicate with the poe.ninja API."""
 from .currency import CurrencyType, get_currency_base_types
-from .item import MiscItemType, UniqueFilter, get_misc_base_types, get_unique_base_types
+from .item import MiscItemType, get_misc_base_types, get_unique_base_types