+++ changelog.md
@@ -1,5 +1,6 @@
 # 1.4.2
-* Fixed a bug where getting text values from a line with `=` or empty operator yielded the opposite result of what was intended. 
+* Fixed a bug where getting text values from a line with `=` or empty operator yielded the opposite result of what was intended.
+* Fixed a bug where not all lines with the same condition in a block were accounted for if any of them had no values.
 
 # 1.4.1
 * Added **Omens** and **Tattoos** to `.econ`.